Consistently handle TESTOPTS
Right now, if you run 'rails test' with TESTOPTS defined in your env
it will be ignored. However, if you run 'rake test' the TESTOPTS variable
will be parsed. This is nnot intuitive and took me awhile to get to the bottom of.

This change eliminates any difference between rails test and rake test.